Revise tag cleanup schemes for better automated pipelines
We've made some changes to the tag deletion recently, and I wanted to make things a bit better, if possible.

Rather than hardcoding certain tagging patterns for optional deletion, just fully parameterize the `git tag -l` pattern that will be cleaned up.

Also, add the same functionality to the `dev-promote-prod` job, since it was already doing some (previously hardcoded) optional tag cleanup.

The goal here is to have orb repos where the only tags that remain, are those following a `vX.Y.Z` SemVer pattern, which is the default published when using `dev-promote-prod` with `publish-version-tag` set to `true`.

The way to achieve this is specified in the default values for `cleanup-tags-pattern` in both jobs in which it now appears:

- In `trigger-integration-workflow`, the default will cleanup all tags starting with the string `integration`, intended to be used only when `trigger-integration-workflow` is run on a commit to master, e.g., when its `tag` parameter is set to `master`.
- In `dev-promote-prod`, the default will cleanup all tags starting with the string `master`, and will only run _after_ a successful orb deployment.

And, as a reminder, these cleanup schemes will _only_ get used when `cleanup-tags` is set to `true` (its default value is `false`).