Adds promote command, updates promote job
This would be a [semver:major] change I think.

Rather than hard-coding the major,minor, patch increment, it will be
deciphered by the commit message.

This is major change because users will get a failure message if they do
not include that in the commit to default branch.

Optionally the prod version can be published by a bot back to Gurhub PR