feat: use load to slurp YAML file
This change leverages the load function from yq.
It allows us to load the full content from a YAML file.
This way, we do not need to cat the YAML file content in a prior step.

In addition, I believe this can avoid the 'Argument list too long' yq error.